Script to help identify NML entries added since last release that are not described in the README.namelist
TYPE: no impact

KEYWORDS: README.namelist, script

SOURCE: internal

DESCRIPTION OF CHANGES:
Script that compares the ARW (non DA, non Chem) registry files with the versions from the last release. Any new additions to the rconfig entries that are not described in the README.namelist file are flagged. Any entirely new registry files are just reported (as not every variable has to be in the README.namelist).

LIST OF MODIFIED FILES:
A       tools/updates_in_registry.csh

TESTS CONDUCTED: Explicitly state if a WTF and or other tests were run, or are pending. For more complicated changes please be explicit!
Seems to work pretty good.