Check for small dt, nint(dt) identically = 0, bad MOD comparison in diags
TYPE: bug fix

KEYWORDS: MOD, diags, timestep, dt

SOURCE: problem found by Brian Reen (US Army ARL)

DESCRIPTION OF CHANGES:
For the test to determine whether or not to do some diagnostic processing, if the time step is small, do the diagnostics at each step. Already the logic was set up so that the the print frequency was identical to the time step for the diag_print option set to 1. Allow this to continue even if the time step is small.

LIST OF MODIFIED FILES:
M       phys/module_diag_misc.F

TESTS CONDUCTED: Explicitly state if a WTF and or other tests were run, or are pending. For more complicated changes please be explicit!
- [ ] Testing to see if it works, not yet
- [ ] Reggie 3.07, not yet