!!!NOT FOR v3.9!!! wrf_io.F90, index 0 invalid
TYPE: bug fix

KEYWORDS: IO, netcdf, wrf_io.F90, index

SOURCE: internal

DESCRIPTION OF CHANGES:
When using -D, the ndown program fails instantly with an index of zero, where the minimum allowed index is one. The real and ideal pre-processors, and the WRF model do not trip up in this case (do not get into this part of the IF test). The ndown program does fail.

With the fix, the results are identical between two cases:
1. Run ndown without -D
2. Run ndown with -D and also with the I/O mod

The change is to put a MAX(1,index) as the array index.

LIST OF MODIFIED FILES:
M       external/io_netcdf/wrf_io.F90

TESTS CONDUCTED:
1. Works with real, WRF, and ndown.
2. Reggie not done yet