TYPE: bug fix (#239)
KEYWORDS: KPP Makefile

SOURCE: internal

DESCRIPTION OF CHANGES:

Altered argument list to linker command so that the libraries are a contiguous
group at the end of the command.  Before this change the object files where
in the middle of the library specifications.

LIST OF MODIFIED FILES:

M	chem/KPP/kpp/kpp-2.1/src/Makefile

TESTS CONDUCTED:

WTF reg test, version 3.07, has been almost successfully completed.
The em_chem suite for the PGI compiler reported a compilation failure.
The em_chem suite does not compile and use kpp.  The compilation
failure was due to missing components in the wrf.exe link phase
and the reason is unknown.