Remove undefined/unnecessary variable from OBSPROC routines #259
TYPE: bug fix

KEYWORDS: WRFDA, OBSPROC, debug, intel

SOURCE: internal

DESCRIPTION OF CHANGES:
Tests on local linux machine result in failure for one obsproc case when compiled with the intel compiler in debug mode. This is due to variable "nvalids", which is used without being defined in some subroutines within the file var/obsproc/src/module_write.F90 (it is defined in the parent subroutine, but not explicitly passed to the child subroutine). Since this variable is not actually used within any of these subroutines in this file, the easiest solution is just to remove it altogether.

LIST OF MODIFIED FILES:
M var/obsproc/src/module_write.F90

TESTS CONDUCTED: WRFDA Regtest now passes this one particular test for intel on local linux machine in debug mode. Other regtests are unaffected as they never reported this error in the first place.