Add a github pull request template (#256)
TYPE: text only

KEYWORDS: github, pull request template

SOURCE: internal

DESCRIPTION OF CHANGES: I recently learned that Github supports the capability to store a pull request template message as a hidden file in the source code.

Currently the template for pull request messages (which become the git log messages we see after the pull request is moved) is stored in the source code under tools/commit_form.txt. This will still be kept for reference, but now a condensed form of this message (with a link to further information) will automatically appear in the text box when you go to open a new pull request.

LIST OF MODIFIED FILES:
A .github/PULL_REQUEST_TEMPLATE

TESTS CONDUCTED: Added this template to my fork and attempted to open a new pull request to the master of my fork, confirmed the expected behavior that the text appeared automatically in the text box!