Make lanczos eigenpair filename a namelist option #267
TYPE: enhancement

KEYWORDS: WRFDA, FSO, lanczos, namelist

SOURCE: internal

DESCRIPTION OF CHANGES: When activating the "lanczos" minimization algorithm you have the option to write the calculated lanczos eigenpairs for future reading (this capability is used in the FSO system). Currently, when that option is activated, these eigenpairs are written by each processor according to the filename pattern "../lanczos_eigenpairs.####". Notice the ".." at the beginning of the filename; this means that the files will be written in the directory ABOVE the current working directory. While this may make sense for the FSO system (which is guided by a number of script in a strict directory hierarchy), running this ability stand-alone (which I am currently doing for debugging purposes) is extremely inconvenient.

This change adds the ability to specify the filename pattern as a namelist option (I say "pattern" because the files will still be appended with the processor number). Because the FSO system requires this pattern, I will keep the default filename identical for back-compatibility.

LIST OF MODIFIED FILES:
M Registry/registry.var
M var/da/da_minimisation/da_lanczos_io.inc
M var/da/da_minimisation/da_minimisation.f90

TESTS CONDUCTED: Ran stand-alone test with lanczos algorithm and filename behavior is as expected. Ran WRFDA regtest for GNU and Intel: no changes as expected.