set shdfac=0 for 25-27 categories only if it is USGS or ISURBAN=1 (#293)
TYPE: bug fix

KEYWORDS: variable shdfac, 25-27 special categories USGS

SOURCE: internal

DESCRIPTION OF CHANGES:

The variable SHDFAC is set to 0 for categories 25 - 27. But this is only true if the landuse type is USGS. For other land types, such as the 40 category NLCD, this is not correct. The fix checks if ISURBAN is 1, and if so, set these SHDFAC for these three categories to 0. This is ok so far, as most other land types has ISURBAN = 13.

LIST OF MODIFIED FILES:

modified: phys/module_sf_noahdrv.F

TESTS CONDUCTED: Passed combined WTF with other changes.