Fix for Ferrier microphysics to work with GFDL radiation in ARW (#289)
TYPE: bug fix

KEYWORDS: Ferrier microphysics, GFDL radiation

SOURCE: internal

DESCRIPTION OF CHANGES: ARW code would stop in the radiation driver if Ferrier is chosen with GFDL because QS is not present. This is because a while back Ferrier was switched to use QI instead of QS. Note NMM lists both in its package, so it runs. We now allow GFDL radiation to run if either QI or QS is present. Note that mp_physics=95 uses QS. Other Ferrier options use QI.

LIST OF MODIFIED FILES:
M phys/module_radiation_driver.F

TESTS CONDUCTED:
Mod was tested by wrfhelp for user problem.
WTF passed for all except a single NMM test with DFI which shows NaNs in the QS field for some compilers at time zero: these NaNs do not seem to impact results in other fields or other times (forecast after time zero is bit-for-bit identical with previous runs). Since this is an extremely minor problem that impacts very few if any users, the failing test will be removed from the WTF for now and an issue opened to be fixed at a later.