Update GEN_BE to allow arbitrary intervals between valid times for NMC method (#307)
TYPE: enhancement

KEYWORDS: WRFDA, GEN_BE, scripts only

SOURCE: internal

DESCRIPTION OF CHANGES: GEN_BE uses the NMC method (comparing forecasts of different lead times valid at the same time) to generate a background error statistics file for the user's specific domain. Currently, due to the way GEN_BE scripts are set up, the lead time of the short forecast must be the same as the interval between forecasts. For example, for NMC 24-12 (comparing a 24-hour forecast to a 12-hour forecast valid at the same time), the interval between forecasts must be 12 hours. For NMC 36-24, the interval between forecasts must be 24 hours. This is an artificial limitation that has no real scientific reason for existing.

This simple change separates the variable INTERVAL so that it now only controls the interval in hours between forecasts. There is now a new variable FCST_RANGE2 (actually an old variable which was unused) which is the shorter forecast time in hours. The change is made in a way that preserves the old behavior if FCST_RANGE2 is not set.

LIST OF MODIFIED FILES:
M var/scripts/gen_be/gen_be_set_defaults.ksh
M var/scripts/gen_be/gen_be_stage0_wrf.ksh
M var/scripts/gen_be/gen_be_wrapper.ksh

TESTS CONDUCTED:
WRFDA regression test passed, old script with same settings gives identical results. Designed an experiment in which GEN_BE was run for NMC24-12, once with an interval of 12 hours, and again with an interval of 36 hours. For the latter experiment, I copied the appropriate data from the 12-hour experiment so that, if the capability was working as expected, the results would be bit-for-bit identical, and they were. Also ran normal experiments with intervals of 6 and 24 hours to ensure be.dat was successfully created and that there were no obviously wrong results.

WTF Not run, as only GEN_BE scripts were changed.