Ensure global domain has polar=.true. in &bdy_control (#302)
TYPE: bug fix

KEYWORDS: global, polar

SOURCE: internal

DESCRIPTION OF CHANGES:

( ( if the domain is the entire globe ) and
( if the user has not set the &bdy_control flag polar=.true. ) ) then
There are troubles, as in the user probably just forgot or did not even know to set the lateral BC switch that specifies turning on the polar filters.

Ergo, stop.

Also, in the real program, remove the old logic for fixing map factors. The chance that someone is using a new version of WRF with SI input is mighty low. Make the error comments to the user match what we are really testing on inside the code.

To determine if the domain is the size of the globe, a couple of constants are used from module_model_constants. That requires an additional USE statement in check_a_mundo, and also a new dependency in main/depend.common.

LIST OF MODIFIED FILES:

M dyn_em/module_initialize_real.F
M share/module_check_a_mundo.F
M main/depend.common

TESTS CONDUCTED:

 Reggie complete - failures seem unrelated
 Global domain, without polar=.true., fails in a controlled fashion
 
&bdy_control
 periodic_x                          = .true., .false.,.false.,
 nested                              = .false., .true., .true.,

 spec_bdy_width                      = 5,
 spec_zone                           = 1,
 relax_zone                          = 4,
 specified                           = .true., .false.,.false.,

 /
 polar                               = .true. ,.false.,.false.,


-------------- FATAL CALLED ---------------
FATAL CALLED FROM FILE:  <stdin>  LINE:     538
Maybe this is a global domain, but the polar flag was not set in the namelist.