Adds Mosaic ver. 2.0 in the chemistry with a selectable namelist option
This commit adds Mosaic version 2 in the chemistry codes. This version
has an updated version of Mosaic with all the bug fixes found after the
initial porting of Mosaic in WRF. Other intent was to ensure that
Mosiac codes used across different models (WRF, CESM, E3SM etc.) are
consistent.