non-zero DTBC in IC file when using adaptive time step: model "completes" immediately
TYPE: bug fix

KEYWORDS: Adaptive time step

SOURCE: Internal

DESCRIPTION OF CHANGES:
Problem: When the variable DTBC is set to a non-zero value in wrfinput, the model exits with "SUCCESS COMPLETE WRF" but it actually didn't start integration. One possible way for this to occur is to have the WRF model use the output option for an input file (mainly used for DA).

There are two solutions, and both are reasonable:

Check if the offending input file is both ( an input file .and. this is domain 1). ALWAYS in that case, set the DTBC variable to zero.
The DTBC variable is computed internally within the WRF model. It should not be input at all. The "i" modifier is removed in the Registry file to remove the input stream for variable 'DTBC'
LIST OF MODIFIED FILES
M share/input_wrf.F
M Registry/Registry.EM_COMMON

TESTS CONDUCTED:

 regresssion test passed
 a separate case with non-zero DTBC in wrfinput is run with and without the modification