Updated to remove most mpas ifdefs
Because the mpas-specific variables are optional, it is cleaner to remove the surrounding mpas ifdefs, and use the if(present(....)) tests for the optional variables; therefore I implemented this.  One mpas-specific ifdef remains because we aren't sure what the calculations are used for and we don't want to mess with it.