Allow non-scalar array 4d arrays to be properly shifted for moving nests
The code to handle shifts of a moving nest in the x- or y-directions previously
assumed that any array with four dimensions whose first and third dimensions
were X and Y, or whose first and second dimensions were X and Y, was a scalar
array. The logic for shifting scalar arrays relies on the 'num_table' member
of the 'fieldlist' type (defined in module_domain_type.F) to be allocated;
however, this is not the case for non-scalar array 4d arrays.

Now, when shifting 4d arrays, we first check whether the 'scalar_array' member
of a 'fieldlist' type is true before applying scalar array-specific logic;
otherwise, the two non-horizontal dimensions (i.e., the dimensions other than
X and Y) of a 4d array are simply shifted as a whole.

Note that, following the logic in lines 352-355 of gen_allocs.c, we are
guaranteed that 'num_table' is associated only if 'scalar_array' is true.