BF of compile script: -j n trips "unrecognized option" exit
TYPE: bug fix

KEYWORDS: compile, -j

SOURCE: Found by Stacy Walters (ACOM), internal

DESCRIPTION OF CHANGES:
The csh for loop used a "shift" to find the argument after the "-j" command line flag. That "shift" command does not actually modify the for loop processing, and the (in the example of "-j 6"), the "6" is seen as an unrecognized argument. If the "-j" argument is found a flag is set to skip over the next argument.

LIST OF MODIFIED FILES:
M   compile

TESTS CONDUCTED:
 - [x] With this mod the code builds.
 - [x] Without this mods, the compile script stops