Add CODEOWNERS to .github directory
TYPE: new feature

KEYWORDS: CODEOWNERS

SOURCE: internal

DESCRIPTION OF CHANGES:
This is the start of a list for shared and individual responsibility for files and directories. When a PR is issued that modifies one of the known files or files within a known directory (and required reviews are enabled), then an automatic request for review is issued.

More info:
```
https://help.github.com/articles/about-codeowners/
```

LIST OF MODIFIED FILES:
M	.guthub/CODEOWNERS

TESTS CONDUCTED:
 - [ ] Going to try out some fake PRs to see what happens.