Fix excessive "outside of domain" prints when assimilating CWB radar mosaic data.
For CWB's radar data assimilation, the radar data are pre-processed
to be on mass points of the model grid.
When WRFDA reads in the radar mosaic data on model mass points,
it prints out a lot of "Report is outside of domain" messages when
print_detail_radar=.true.

Those "outside of domain" obs appear to be along the domain boundaries.

When calculating grid x/y from lat/lon (by calling da_llxy),
the output x/y are floating numbers and they do not match exactly the grid index.
For example, by giving the lat/lon of (450,1),
the calculated x is 450.000183 and y is 0.999511719.
A few more examples are listed below:
  (450,   2) :  (450.000183,    2.00000000)
  (450,   3) :  (450.000153,    3.00000000)
  (450,   4) :  (450.000153,    3.99951172)
  (450,   5) :  (450.000092,    5.00048828)
  (450,   6) :  (450.000061,    6.00000000)
  (446, 450) :  (446.000580,  450.000000)
  (447, 450) :  (446.999146,  449.999512)
  (448, 450) :  (448.000671,  449.999512)
  (449, 450) :  (448.999268,  450.000000)
  (450, 450) :  (450.000793,  450.000000)

By removing the "=" sign in line 52 and line 53 of var/da/da_tools/da_llxy.inc,
the "outside of domain" prints can be reduced.

Another solution to reducing run-time output is to set print_detail_radar=.false.
for operational runs.

modified:   var/da/da_tools/da_llxy.inc