compile target "wrf" is now available again (#453)
TYPE: enhancement

KEYWORDS: compile, wrf

SOURCE: internal

DESCRIPTION OF CHANGES:
Add in the target "wrf" in the top if test in the compile script. The test is placed after the "wrfplus" target. The only executable to be built is wrf.exe, so that is the only executable that is removed.

LIST OF MODIFIED FILES:
M    compile

TESTS CONDUCTED:
 - [x] Without the mod, the build target is not recognized.
 - [x] With the mod, the wrf.exe (and only wrf.exe) is rebuilt.