modified to prevent floating point error with WETCAN calculation (#456)
TYPE: bug fix

KEYWORDS: WETCAN, CN, positive_definite, morrison, RUC

SOURCE: internal

DESCRIPTION OF CHANGES: When using the Morrison microphysics scheme, along with the RUC LSM, an fatal error is produced: "Crash in surface energy budget," or sometimes a floating point error. This is reproducible with a basic (near-default) namelist set-up. Traced back to module_sf_ruclsm.F where WETCAN is assigned a value that can be non-real, depending on values of CST and SAT. Since CN is always equal to 0.5, if CST or SAT are negative, then we are taking the square root of a negative number. Corrected so that this is always positive.

LIST OF MODIFIED FILES:
M module_sf_ruclsm.F

TESTS CONDUCTED: wtf passes, tested modified code to make sure the error is gone