TEMPORARY bug fix for build after splitting sea level pressure from module_diag_afwa.F
TYPE: bug fix

KEYWORDS: temporary, build

SOURCE: internal

DESCRIPTION OF CHANGES:

The better working solution, though not tested as well, is to have the depend.common have this structure:
```
a.o: b.o
b.o: c.o
```
instead of the current
```
c.o:
<tab>explicit rules to build c.o
a.o: c.o
b.o: c.o
```

LIST OF MODIFIED FILES:
M	main/depend.common

TESTS CONDUCTED: Explicitly state if a WTF and or other tests were run, or are pending. For more complicated changes please be explicit!
In the main/depend.common, a working method to build the code reproducibly is to insert the build rules in the depend.common file. I tested this with all combinations of the following:
1. compiler: Intel, GNU, PGI
2. parallel make: -j = 1, 2, 4, 6, 8
3. cheyenne and local Mac
4. optimized, -d, -D
5. manually built and through the entire regression test
For most of the above I ran multiple instances.