Auto detect if NETCDF4 with compression is not available, fallback to NETCDF classic
TYPE: enhancement

KEYWORDS: NETCDF, compression, classic

SOURCE: internal

DESCRIPTION OF CHANGES:
Use an if test in the configure script. If the attempted NETCDF4 build fails on a small test, then change the NETCDF format to classic, and continue the rest of the build.

LIST OF MODIFIED FILES:
M	   configure

TESTS CONDUCTED:
1. Original code, when requesting NETCDF4 with compression (default), and the libraries do not exist.
```
************************** W A R N I N G ************************************
NETCDF4 IO features are requested, but this installation of NetCDF
  /usr/local/netcdf-3.6.3-gfortran-6.2.0
DOES NOT support these IO features.

Please make sure NETCDF version is 4.1.3 or later and was built with
--enable-netcdf4

OR set NETCDF_classic variable
   bash/ksh : export NETCDF_classic=1
      echo  csh : setenv NETCDF_classic 1

Then re-run this configure script

!!! configure.wrf has been REMOVED !!!

*****************************************************************************
```

2. New code when asking for NETCDF4 with compression, and the libraries do not exist
```

This build of WRF will use classic (non-compressed) NETCDF format
```

3. New code when asking for NETCDF4 with compression, and the libraries DO exist
```
Something
```