Add handy build info after compile of em_fire
TYPE: enhancement

KEYWORDS: em_fire, Makefile

SOURCE: internal

DESCRIPTION OF CHANGES:
The ARW builds all have a nice message at the end of the build. If the build worked, you get a message concerning how long the build required and a listing of the generated executables. If the builds was not successful, there is a message that problems were encountered. While testing the separated ideal initialization modules (tropical cyclone, held suarez, fire, scm), the em_fire build did not produce the nice message, so that is now rectified.

LIST OF MODIFIED FILES:
M	   Makefile

TESTS CONDUCTED:
1. A couple of dozen builds with em_fire during some recent development demonstrate that this new code is OK.
2. Without the mods, there is no final message after a build about the status of the executables.