Add CODEOWNERS file to .github directory
TYPE: new feature

KEYWORDS: CODEOWNERS

SOURCE: internal

DESCRIPTION OF CHANGES:
This is the start of a list for shared and individual responsibility
for files and directories. When a PR is issued that modifies one of
the known files or files within a known directory (and required
reviews are enabled on that branch), then an automatic request for
review is issued to the code owners.

CODEOWNERS files is linked to the GitHub via the following teams:
   * physics
   * da
   * chem
   * hwrf_dev_team
   * arw_dev_team
   * infrastructure
   * global_owner
   * fire

More info:
```
https://help.github.com/articles/about-codeowners/
```

LIST OF MODIFIED FILES:
M	     .github/CODEOWNERS

TESTS CONDUCTED:
 - [x] Required code owner to review before merge was permitted.