Bug fix: cray, do loop index is real, module_mp_morr_two_moment_aero.F
TYPE: bug fix

KEYWORDS: cray, do loop, real

SOURCE: Alex ??? (US Naval Academy)

DESCRIPTION OF CHANGES:
An index for a loop counter and also used for an array index was declared as
a real value. This is changed to an integer. There is no other place than the
do loop or the array index that this value was used.

LIST OF MODIFIED FILES:
M   module_mp_morr_two_moment_aero.F

TESTS CONDUCTED:
1. Code does not compile without this mod.
2. With this mod, the code compiles.