MYNN PBL shcu combo bug #680
TYPE: bug fix

KEYWORDS: mynn, bug, shcu, ishallow, bl_mynn_edmf, check_a_mundo, fatal

SOURCE: internal

DESCRIPTION OF CHANGES:
Problem:
The MYNN PBL has the EDMF (bl_mynn_edmf=1) option turned on by default. In check_a_mundo when the bl_mynn_edmf=1, (even if the user specifically asks for the shallow cumulus option to be activated (sh_cu=1)), the shallow cumulus scheme is simply turned off.

Solution:
The check_a_mundo test is modified to give a fatal error when bl_mynn_edmf=1 AND sh_cu=1 so that a user will not overlook this message in the log file.

ISSUE:
This addresses the issue referenced in #617 "MYNN PBL can't be used with SHCU"
closes:#617

LIST OF MODIFIED FILES:
M share/module_check_a_mundo.F

TESTS CONDUCTED: Verified that code builds with modification. Verified that when using the MYNN PBL, and turning on ishallow, shcu_physics, or both, the real.exe or wrf.exe program stops with the intended fatal error message.

RELEASE NOTE: A check was implemented to stop the real or wrf program if bl_mynn_edmf (which is set to 1 by default in the MYNN scheme) is being used in combination with either the ishallow or shcu_physics options turned on. A message informs the user that bl_mynn_edmf cannot be used with the other 2 options and the namelist must be modified.