Include share and frame dirs in CODEOWNERS files
TYPE: bug fix

KEYWORDS: CODEOWNERS, share, frame

SOURCE: internal

DESCRIPTION OF CHANGES:
Both the frame and the share directories were accidentally excluded
from being assigned in the CODEOWNERS file. This puts all of those
files and directories under the default control of global-owner.
Since these are infrastructure directories, that is the group that
should have ownership.

LIST OF MODIFIED FILES:
M	.github/CODEOWNERS

TESTS CONDUCTED:
 [x] Default CODEOWNER was notified of required review.

RELEASE NOTE: These mods are for the internal workings of the merge
procedure to the develop branch, so no need to mention this PR in the
release notes.