Add 1D domain check to reduce time for processing prepbufr files in in da_read_obs_bufr
In my experience, this modification reduced the time to process a single BUFR file
from 13 minutes to less than 20 seconds.  There is some memory overhead, but it
does not appear to be a limiting factor even with 36 processors per node.  Additional
distribution of work could be accomplished if there prove to be memory limits, which
is given as a NOTE in da_read_obs_bufr.inc.

 Changes to be committed:
	modified:   da_obs_io/da_obs_io.f90
        modified:   da_obs_io/da_read_obs_bufr.inc
	modified:   da_tools/da_llxy_1d.inc