Bug fix for the adjustment of urban type in REAL program when NLCD dataset is used (#702)
TYPE: bug fix

KEYWORDS: urban, land use, MODIS, REAL, NLCD

SOURCE: internal

DESCRIPTION OF CHANGES:
Problem:
When NLCD and MODIS are used together, the land use type will be adjusted for all grid cells where the urban fraction is greater than 0.5 in the real program. Due to an incorrect application of DeMorgan's Law, all of the urban land use categories became category 13. While 13 for MODIS is indeed urban, 13 is not the NLCD high-resolution urban categories: 24, 25, 26. These NLCD urban categories should be preferentially used if available.

Solution:
The logic is supposed to be:
if ( the grid point is urban (based on percentage) ) and
( the grid cell is not one of the special NLCD points ) then
reset this to the MODIS urban category.

if (urban>0.5)
   if ((cat != 24) AND cat(!=25) AND cat(!=26))
      cat = 13
Note that with our incorrect implementation of the logic, the if test above is ALWAYS true: we always reset the urban points to the MODIS value. Here is the line that we need:

if ((cat != 24) OR cat(!=25) OR cat(!=26))
The incorrect logic in the IF test means that all urban points (based on percentage) become category 13, which is the MODIS urban category. The correct logic excludes urban points that are identified as NLCD urban points.

This bug fix retains the correct urban type defined in NLCD.

LIST OF MODIFIED FILES:
M dyn_em/module_initialize_real.F

TESTS CONDUCTED:

Application to a user case confirms the modified code works as expected.