Protect Goddard radiation from real*8 builds
The input file for Goddard allocates space for the look-up tables.
Either a separate look-up table for real*8 values is required or
logic to support reading in real*4 (regardless of the native
floating point precision). Because there is some additional
infrastructure involved (defined types), before modifying
code in a few places, we'll find out what the NASA developers
want to do.

An ifdef inside of check_a_mundo detects if this is real*8,
and a standard IF test checks for Goddard LW or SW. A controlled
fatal error is included.