Fix -j parallel make option for compile script
TYPE: bug fix

KEYWORDS: compile, -j, parallel

SOURCE: Found by Michael Duda (NCAR), internal fix

DESCRIPTION OF CHANGES:
Problem:
The user is able to specify directly the env variable for a parallel make:
```
setenv J = "-j 4"
```
The user is supposed to be able to also specify this on the command line for
the compile script:
```
compile em_real -j 4
```
However, depending on order of arguments, the number of parallel builds is
incorrectly used in the compile script.

Solution:
The arguments, `$argv` to the compile command can either be processed sequentially
via the `$argv` list and a `shift` command, or the entire `$argv` can be used
in a `foreach` loop. The script intermingled these. The new approach keeps the
`foreach` loop and removes the assignment of the argument for `-j` via the incorrect
`shift $argv`.

ISSUE:
Fixes #778 "compile -j X not working"

LIST OF MODIFIED FILES: list of changed files (use `git diff --name-status master` to get formatted list)
M   compile

TESTS CONDUCTED:
 - [x] If `-j 5 is first argument, everything is OK:
```
> compile -j 5 em_real
next arg is 5
```
 - [x] If `-j 5 is middle argument, everything is OK:
```
> compile io -j 5 em_real
next arg is 5
```
 - [x] If `-j 5` is last argument, everything is OK:
```
> compile em_real -j 5
next arg is 5
```