BF: AGL interpolation incorrect for zld diagnostics
TYPE: bug fix

KEYWORDS: AGL, ZLD, diagnostics

SOURCE: internal

DESCRIPTION OF CHANGES:
Problem:
The height-level diagnostic routine has an option that the user may specify
either the height above mean sea level or the height above ground level (AGL).
If a user chooses 50 m AGL, the height field should be almost exactly 50 m
over the ocean, and almost exactly 50 m greater than the topo over land.
This is not what was happening.

Solution:
The height trapping levels (in the WRF model, (ph+phb)/g) should be as identical
to the WRF model levels (they were not). The AGL level to interpolate to, should be
the requested AGL level + the topo value (it was not).

LIST OF MODIFIED FILES:
M   phys/module_diag_zld.F

TESTS CONDUCTED:
This is purely a diagnostic, and does not impact the model simulation. A valid test
is to look at the height field with a know topography elevation. The AGL height field
should be the AGL level + the topo value at that grid point.

RELEASE NOTE:
Since the height level interpolation routine was introduced, it has used an incorrect
formualtion for the AGL computations.