Replace "use mpi" with "include mpif.h"
Most of the WRF model used the mpif.h file for the MPI types, externals
and constants. A few places have added "use mpi" instead.

In addition to consistency, this makes default installed implmentations
(yum, apt-get, brew) of mpich and openmpi able to be used when ALSO
installing a newer gfortran compiler.