logical specified should be defined before it is used (#897)
TYPE: bug fix

KEYWORDS: diff_6th_opt, specified

SOURCE: Robert Rozumalski (COMET)

DESCRIPTION OF CHANGES:
When a restart problem was fixed in 4.0.3 (Dec 18, 2018), an error was introduced. It used a logical
variable 'specified' in subroutine sixth_order_diffusion, but it was not defined before it was used. This
caused model failure in the case reported by Rob. This PR fixes the problem.

ISSUE:
Fixes #895 "module_big_step_utilities_em.F - logical variable "specified" not assigned"

LIST OF MODIFIED FILES:
M dyn_em/module_big_step_utilities_em.F

TESTS CONDUCTED:
Compiled and verified it fixed the user's problem.

RELEASE NOTE:
Fixed a bug associated with option diff_6th_opt. When a restart problem was fixed in 4.0.3 (Dec 18, 2018), an error was introduced. It used a logical variable 'specified' in subroutine sixth_order_diffusion, but it was not defined before it was used. This caused model failure in a user's case.