add two , but it does not affect anything (#898)
TYPE: text only

KEYWORDS: missing "," LANDUSE.TBL

SOURCE: Rob Gilliam (EPA)

DESCRIPTION OF CHANGES:
There are two commas ',' missing from LANDUSE.TBL.

As was verified this does not affect model results at all.
Also, this is not an error that a different compiler would find. This is not an error.
This is a cosmetic fix to make the text-based file self-consistent.
So we consider this as a 'text change' only.

ISSUE:
Fixes #830 "Typo in LANDUSE.TBL"

LIST OF MODIFIED FILES:
M run/LANDUSE.TBL

TESTS CONDUCTED:
Substantial testing was conducted with the reported issue. An extracted subroutine read both the
original and new files. No change in results. The following is taken from the issues page:

There should be no impact of the missing comma. Since this is list directed input, either white
space or a comma would be suitable for parsing the individual fields.

D. pulled out the read portion of the routine that handles this file to verify this assertion and ran
the code for both MODIS and MODIFIED_IGBP_MODIS_NOAH, and with the original
run/LANDUSE.TBL and a corrected file. No diffs in the printed standard out.