Bug fix: traditional fields diags, u and v are 2x too large
TYPE: bug fix

KEYWORDS: diagnostics

SOURCE: Nicolas Baldeck

DESCRIPTION OF CHANGES:
When computing the earth-relative winds from the model's projection-relative winds, the resulting values of umet and vmet are correctly placed in the cell center. However, the averaging to the cell center from the momentum cell faces is missing a "divide by 2".

ISSUE:
Closes the wind portion of the problem note in issue #935 "Bad Wind and Temperature values with diag_nwp2=1"

LIST OF MODIFIED FILES:
modified:   module_diag_trad_fields.F

TESTS CONDUCTED:
 - [ ] None yet

RELEASE NOTE: One of the diagnostic options in WRF is called "traditional fields". One of the available fields is the earth-relative components of the horizontal wind. Those values were too large by a factor of two (due to averaging to the cell center, but neglecting to divide by 2).